Refactored searchFacts further into separate component and passed unit test to render facts
